bug fix in mkSD ("molecule out of box")
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/density-profiles@1289 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ tools/mksd/GlobalStartGeometry.h
@@ -75,7 +75,7 @@ private:
 	//@brief: lengths of a single elementary liquid lattice in three directions
 	double _liqUnit[3];
 	//@brief: number of elementary vapour lattice boxes per direction
-	unsigned _vapUnits[3];
+	long int _vapUnits[3];
 	//@brief: lengths of a single elementary vapour lattice with respect to three directions
 	double _vapUnit[3];
 	//@brief: gross probability of a liquid elementary box to be filled
